Enhancement: Merge Definitions into FixtureFactory::load()
